feat: when NSItemProvider points to a local file, use it as is, no longer creating a webloc fix: unable to share a ProCreate file
