--- Sources/InfomaniakCore/ItemProviderRepresentation/ItemProviderWeblocRepresentation.swift
@@ -1,108 +1,0 @@
-/*
- Infomaniak Mail - iOS App
- Copyright (C) 2022 Infomaniak Network SA
-
- This program is free software: you can redistribute it and/or modify
- it under the terms of the GNU General Public License as published by
- the Free Software Foundation, either version 3 of the License, or
- (at your option) any later version.
-
- This program is distributed in the hope that it will be useful,
- but WITHOUT ANY WARRANTY; without even the implied warranty of
- MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- GNU General Public License for more details.
-
- You should have received a copy of the GNU General Public License
- along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-#if canImport(MobileCoreServices)
-
-import Foundation
-import InfomaniakDI
-
-/// Something that can provide a `Progress` and an async `Result` in order to make a webloc plist from a `NSItemProvider`
-@available(macOS 10.15, iOS 13.0, tvOS 13.0, watchOS 6.0, *)
-public final class ItemProviderWeblocRepresentation: NSObject, ProgressResultable {
-    /// Progress increment size
-    private static let progressStep: Int64 = 1
-
-    /// Number of steps to complete the task
-    private static let totalSteps: Int64 = 2
-
-    /// Something to transform events to a nice `async Result`
-    private let flowToAsync = FlowToAsyncResult<Success>()
-
-    /// Shorthand for default FileManager
-    private let fileManager = FileManager.default
-
-    /// Domain specific errors
-    public enum ErrorDomain: Error, Equatable {
-        case unableToLoadURLForObject
-    }
-
-    public typealias Success = URL
-    public typealias Failure = Error
-
-    public init(from itemProvider: NSItemProvider) throws {
-        progress = Progress(totalUnitCount: Self.totalSteps)
-
-        super.init()
-
-        let completionProgress = Progress(totalUnitCount: Self.progressStep)
-        progress.addChild(completionProgress, withPendingUnitCount: Self.progressStep)
-
-        let loadURLProgress = itemProvider.loadObject(ofClass: URL.self) { [self] path, error in
-            guard error == nil, let path: URL = path else {
-                let error: Error = error ?? ErrorDomain.unableToLoadURLForObject
-                completionProgress.completedUnitCount += Self.progressStep
-                flowToAsync.sendFailure(error)
-                return
-            }
-
-            // Save the URL as a webloc file (plist)
-            let content = ["URL": path.absoluteString]
-
-            do {
-                @InjectService var pathProvider: AppGroupPathProvidable
-
-                // Use a unique folder to prevent collisions
-                let tmpDirectoryURL = pathProvider.tmpDirectoryURL
-                    .appendingPathComponent(UUID().uuidString, isDirectory: true)
-                try fileManager.createDirectory(at: tmpDirectoryURL, withIntermediateDirectories: true)
-
-                let currentName = (path.lastPathComponent as NSString).deletingPathExtension
-                let fileName: String
-                if currentName.isEmpty {
-                    fileName = "\(URL.defaultFileName()).webloc"
-                } else {
-                    fileName = "\(currentName).webloc"
-                }
-
-                let targetURL = tmpDirectoryURL.appendingPathComponent(fileName)
-                let encoder = PropertyListEncoder()
-                let data = try encoder.encode(content)
-                try data.write(to: targetURL)
-
-                completionProgress.completedUnitCount += Self.progressStep
-                flowToAsync.sendSuccess(targetURL)
-            } catch {
-                completionProgress.completedUnitCount += Self.progressStep
-                flowToAsync.sendFailure(error)
-            }
-        }
-        progress.addChild(loadURLProgress, withPendingUnitCount: Self.progressStep)
-    }
-
-    // MARK: ProgressResultable
-
-    public var progress: Progress
-
-    public var result: Result<URL, Error> {
-        get async {
-            return await flowToAsync.result
-        }
-    }
-}
-
-#endif
